Update 2021_Popovic_Tiwanaku.janno
while reviewing other package I noticed that it is possible to put two different strategies in the Capture_type field. For some reason I wasn't able to do this previously. So now for two samples for which the genotype was obtained from merged shotgun-arbor libraries I updated this field to ArborPrimePlus;Shotgun